Make TreehouseApp an interface
